third_party/WebKit/LayoutTests/webmidi/requestmidiaccess-in-detached-frame.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests navigator.requestMIDIAccess fails in a detached frame.");

window.jsTestIsAsync = true;
iframe = document.createElement('iframe');
document.body.appendChild(iframe);
contentNavigator = iframe.contentWindow.navigator;
document.body.removeChild(iframe);

// set testRunner configuration so that requestMIDIAccess returns "InvalidStateError".
contentNavigator.requestMIDIAccess().then(function() {
    testFailed("requestMIDIAccess() does not fail unexpectedly.");
    finishJSTest();
}, function() {
    errorName = error.name;
    shouldBe("errorName", "'InvalidStateError'");
    testPassed("requestMIDIAccess() fail as expected.");
    finishJSTest();
});
// FIXME: errorCallback should be called, but due to the Blink Promise impelementation the
// returned Promise never be resolved nor rejected.
finishJSTest();

</script>
</body>
</html>
